requirements.txt
python-docx==1.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1
tqdm==4.66.1
reportlab==4.0.7